refactor for clarity and hygiene. should reduce redundancy and help with scaling out test cases.
